Fixed a number of type mismatch warnings in tests

+++ Realm/Tests/ArrayTests.m
@@ -70,7 +70,7 @@
        
     RLMArray *result = [realm objects:[AggregateObject className] where:[NSPredicate predicateWithFormat:@"intCol < %i", 100]];
     
-    XCTAssertEqual(result.count, 10, @"10 objects added");
+    XCTAssertEqual(result.count, (NSUInteger)10, @"10 objects added");
     
     int totalSum = 0;
